# sql: don't enable "kv tracing" for SHOW TRACE FOR <stmt>