--- a/pkg/sql/logictest/testdata/logic_test/show_trace
+++ b/pkg/sql/logictest/testdata/logic_test/show_trace
@@ -31,7 +31,6 @@ SELECT span, message, operation FROM [SHOW TRACE FOR SELECT 1]
 (0,0)  === SPAN START: sql txn implicit ===  sql txn implicit
 (0,1)  === SPAN START: starting plan ===     starting plan
 (0,2)  === SPAN START: consuming rows ===    consuming rows
-(0,2)  output row: [1]                       consuming rows
 (0,2)  plan completed execution              consuming rows
 (0,2)  resources released, stopping trace    consuming rows
 
@@ -44,7 +43,6 @@ SELECT span, message, operation FROM [SHOW TRACE FOR EXECUTE x]
 (0,0)  === SPAN START: sql txn implicit ===  sql txn implicit
 (0,1)  === SPAN START: starting plan ===     starting plan
 (0,2)  === SPAN START: consuming rows ===    consuming rows
-(0,2)  output row: [1]                       consuming rows
 (0,2)  plan completed execution              consuming rows
 (0,2)  resources released, stopping trace    consuming rows