--- a/pkg/sql/sqlbase/rowwriter.go
+++ b/pkg/sql/sqlbase/rowwriter.go
@@ -250,7 +250,7 @@ func (ri *RowInserter) InsertRow(
 		}
 	}
 
-	if err := ri.Fks.checkAll(ctx, values, traceKV); err != nil {
+	if err := ri.Fks.checkAll(ctx, values); err != nil {
 		return err
 	}
 
@@ -619,12 +619,14 @@ func (ru *RowUpdater) UpdateRow(
 	}
 
 	if rowPrimaryKeyChanged {
-		if err := ru.Fks.checkIdx(ctx, ru.Helper.TableDesc.PrimaryIndex.ID, oldValues, ru.newValues, traceKV); err != nil {
+		if err := ru.Fks.checkIdx(
+			ctx, ru.Helper.TableDesc.PrimaryIndex.ID, oldValues, ru.newValues,
+		); err != nil {
 			return nil, err
 		}
 		for i := range newSecondaryIndexEntries {
 			if !bytes.Equal(newSecondaryIndexEntries[i].Key, secondaryIndexEntries[i].Key) {
-				if err := ru.Fks.checkIdx(ctx, ru.Helper.Indexes[i].ID, oldValues, ru.newValues, traceKV); err != nil {
+				if err := ru.Fks.checkIdx(ctx, ru.Helper.Indexes[i].ID, oldValues, ru.newValues); err != nil {
 					return nil, err
 				}
 			}
@@ -636,7 +638,9 @@ func (ru *RowUpdater) UpdateRow(
 		if err := ru.rd.DeleteRow(ctx, b, oldValues, traceKV); err != nil {
 			return nil, err
 		}
-		if err := ru.ri.InsertRow(ctx, b, ru.newValues, false, traceKV); err != nil {
+		if err := ru.ri.InsertRow(
+			ctx, b, ru.newValues, false /* ignoreConflicts */, traceKV,
+		); err != nil {
 			return nil, err
 		}
 		return ru.newValues, nil
@@ -744,7 +748,7 @@ func (ru *RowUpdater) UpdateRow(
 		secondaryIndexEntry := secondaryIndexEntries[i]
 		var expValue interface{}
 		if !bytes.Equal(newSecondaryIndexEntry.Key, secondaryIndexEntry.Key) {
-			if err := ru.Fks.checkIdx(ctx, ru.Helper.Indexes[i].ID, oldValues, ru.newValues, traceKV); err != nil {
+			if err := ru.Fks.checkIdx(ctx, ru.Helper.Indexes[i].ID, oldValues, ru.newValues); err != nil {
 				return nil, err
 			}
 
@@ -869,7 +873,7 @@ func MakeRowDeleter(
 func (rd *RowDeleter) DeleteRow(
 	ctx context.Context, b *client.Batch, values []parser.Datum, traceKV bool,
 ) error {
-	if err := rd.Fks.checkAll(ctx, values, traceKV); err != nil {
+	if err := rd.Fks.checkAll(ctx, values); err != nil {
 		return err
 	}
 
@@ -891,7 +895,7 @@ func (rd *RowDeleter) DeleteRow(
 	if traceKV {
 		log.VEventf(ctx, 2, "DelRange %s - %s", rd.startKey, rd.endKey)
 	}
-	b.DelRange(&rd.startKey, &rd.endKey, false)
+	b.DelRange(&rd.startKey, &rd.endKey, false /* returnKeys */)
 	rd.startKey, rd.endKey = nil, nil
 
 	return nil
@@ -902,7 +906,7 @@ func (rd *RowDeleter) DeleteRow(
 func (rd *RowDeleter) DeleteIndexRow(
 	ctx context.Context, b *client.Batch, idx *IndexDescriptor, values []parser.Datum, traceKV bool,
 ) error {
-	if err := rd.Fks.checkAll(ctx, values, traceKV); err != nil {
+	if err := rd.Fks.checkAll(ctx, values); err != nil {
 		return err
 	}
 	secondaryIndexEntry, err := EncodeSecondaryIndex(