A "kv trace" is the a set of log messages around SQL<->KV interaction.
The concept is not very well defined at the moment, but it comprises two
categories of messages:
1. Details of KV requests being sent for execution, in particular
pretty-printed scan requests.
2. Per-row messages. In particular, every row that's being returned by
KV for processing by SQL.

These messages are the only ones that are returned as results of
SHOW _KV_ TRACE FOR <stmt>/SESSION.
The situation gets interesting for SHOW TRACE FOR ... (no KV specified).
In this case, before this patch, all the "kv trace" messages were
included in the results. This patch changes that and eliminates the
per-row messages. This is an improvement, per-row messages (as opposed
to per-RPC or higher level) is too verbose for most people that want
traces.

Tactically, a couple of "kv messages" that weren't per row are now
always logged (and hence traced) regardless of whether kv tracing is
enabled or not, and otherwise kv tracing is now no longer enabled for
SHOW TRACE FOR.

This was already the case before this patch, but now this comes to the
foreground: "kv tracing" means one thing externally (as in SHOW KV TRACE
FOR ...), versus internally, in the code, where "kvTracedEnabled" and
similarly-named members refer to "per-row" messages. A round of renames
is in order, but that's left for another day.

Benchstat - no difference
https://gist.github.com/andreimatei/c8ed8f850b59a9eec67b2fb190160108

This patch has no influence on the new index-join benchmark:
```
make bench PKG=./pkg/sql BENCHES=BenchmarkIndexJoin TESTFLAGS='-count 10'
make bench PKG=./pkg/sql BENCHES=BenchmarkIndexJoin TESTFLAGS='-test.benchmem -count 10'

[andrei:~/work/src/github.com/cockroachdb/cockroach] sql-kv-tracing+ ± benchstat /tmp/{before,after}
name         old time/op  new time/op  delta
IndexJoin-8  8.54ms ± 4%  8.54ms ± 1%   ~     (p=0.605 n=9+9)

[andrei:~/work/src/github.com/cockroachdb/cockroach] sql-kv-tracing+ ± benchstat /tmp/{before-mem,after-mem}
name         old time/op    new time/op    delta
IndexJoin-8    8.64ms ± 2%    8.59ms ± 6%    ~     (p=0.243 n=9+10)

name         old alloc/op   new alloc/op   delta
IndexJoin-8    3.62MB ± 0%    3.62MB ± 0%  -0.03%  (p=0.000 n=9+9)

name         old allocs/op  new allocs/op  delta
IndexJoin-8     11.6k ± 0%     11.6k ± 0%    ~     (p=0.314 n=10+10)
```