@andreimatei 2017-10-03
I probably want to cherry-pick this in some 1.1.x release.