@jordanlewis 2017-10-03
Ah I see - this explains why `SHOW KV TRACE` actually currently returns *more* stuff than `SHOW TRACE`, which seems a little unintuitive to me. I like this idea.

I'm surprised that no logic tests change here. I think we should have at least a modicum of test coverage - our traces are actually a very important feature that we can't allow regression in.

---

Reviewed 2 of 2 files at r1.
Review status: 2 of 8 files reviewed at latest revision, 4 unresolved discussions, all commit checks successful.

---

*[pkg/sql/session.go, line 1427 at r2](https://reviewable.io:443/reviews/cockroachdb/cockroach/19008#-KvZE1kRd2tBHp81l8Hi:-KvZE1kSYHwxVYOSELQ-:bfz4vpl) ([raw file](https://github.com/cockroachdb/cockroach/blob/eca98ea2a80df677baa30f90cd5076d8bae76299/pkg/sql/session.go#L1427)):*
> ```Go
> //   verbose messages around the interaction of SQL with KV. Some of the messages
> //   are per-row.
> func (st *SessionTracing) StartTracing(recType tracing.RecordingType, KVTracingEnabled bool) error {
> ```

nit: capitalized argument names seem weird, or at least not "Roman" with respect to our codebase's Rome...

---

*[pkg/sql/trace.go, line 44 at r2](https://reviewable.io:443/reviews/cockroachdb/cockroach/19008#-KvZFVY_8ozvP4QZxsFF:-KvZFVY_8ozvP4QZxsFG:b2yzult) ([raw file](https://github.com/cockroachdb/cockroach/blob/eca98ea2a80df677baa30f90cd5076d8bae76299/pkg/sql/trace.go#L44)):*
> ```Go
> 	// If set, the trace will also include "KV trace" messages - verbose messages
> 	// around the interaction of SQL with KV. Some of the messages are per-row.
> 	KVTracingEnabled bool
> ```

nit: i don't think this should be capitalized since it's not exported. Or is it?

---

*[pkg/sql/trace.go, line 59 at r2](https://reviewable.io:443/reviews/cockroachdb/cockroach/19008#-KvZFZMyue5HLNvNiAkj:-KvZFZMyue5HLNvNiAkk:b-t3slf1) ([raw file](https://github.com/cockroachdb/cockroach/blob/eca98ea2a80df677baa30f90cd5076d8bae76299/pkg/sql/trace.go#L59)):*
> ```Go
> //   verbose messages around the interaction of SQL with KV. Some of the
> //   messages are per-row.
> func (p *planner) makeTraceNode(plan planNode, KVTracingEnabled bool) (planNode, error) {
> ```

nit: caps again.

---

*[pkg/sql/sqlbase/kvfetcher.go, line 227 at r2](https://reviewable.io:443/reviews/cockroachdb/cockroach/19008#-KvZF0ROAeMSnNp1agq7:-KvZF0ROAeMSnNp1agq8:bxwy0u4) ([raw file](https://github.com/cockroachdb/cockroach/blob/eca98ea2a80df677baa30f90cd5076d8bae76299/pkg/sql/sqlbase/kvfetcher.go#L227)):*
> ```Go
> 	}
> 
> 	scanStr := "Scan "
> ```

Hmmm.... not sure if this makes sense. This emits at least 1 Scan event per read operation. Isn't that pretty expensive?

In general it's not obvious to me that we should ever be doing this much unconditional tracing. But I guess that's another fish to fry - we already do a ton of unconditional tracing (including internals like time spent in the command queue). 

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/19008#-:-KvZFrOk7RXvucKvP6Uo:bpk69s4)*
<!-- Sent from Reviewable.io -->