@andreimatei 2017-10-04
> Ah I see - this explains why SHOW KV TRACE actually currently returns more stuff than SHOW TRACE, which seems a little unintuitive to me. I like this idea.

Well, no; before this patch, `SHOW TRACE FOR <stmt>` would include all the messages from `SHOW KV TRACE FOR <stmt>`, so the former should always be a superset. This patch makes the two uncomparable: `SHOW KV TRACE FOR <stmt>` will return a restricted set of messages, but include some per-row ones. The per-row ones are excluded from `SHOW TRACE FOR <session>`. Similarly with the recording done for `SET tracing cluster` vs `SET tracing kv`.
If this is not what you currently see, show me.

"KV tracing" is generally a concept that I have mixed feelings on (I was not involved in creating it). I don't have a clear opinion on whether it deserves to exists and, if it does, exactly what semantics it should have. The thing that I'm interested in and that this patch does, however, is for non-kv tracing to not include per-row messages.


> I'm surprised that no logic tests change here. I think we should have at least a modicum of test coverage - our traces are actually a very important feature that we can't allow regression in.

What test did you have in mind exactly? There is test coverage for tracing, kv and regular. However, the tests are never about checking exactly what a trace looks like; that'd be unwieldy. This PR eliminates a few messages from non-kv traces. How did you expect that to have been tested?
I don't see a particular need for extra testing, but if you have a reasonable idea, I'm opened.

---

Review status: 2 of 8 files reviewed at latest revision, 4 unresolved discussions, all commit checks successful.

---

*[pkg/sql/session.go, line 1427 at r2](https://reviewable.io:443/reviews/cockroachdb/cockroach/19008#-KvZE1kRd2tBHp81l8Hi:-Kv_HF9s3eGhNNy17Z0t:b-896fix) ([raw file](https://github.com/cockroachdb/cockroach/blob/eca98ea2a80df677baa30f90cd5076d8bae76299/pkg/sql/session.go#L1427)):*
<details><summary><i>Previously, jordanlewis (Jordan Lewis) wrote…</i></summary><blockquote>

nit: capitalized argument names seem weird, or at least not "Roman" with respect to our codebase's Rome...
</blockquote></details>

Done.

---

*[pkg/sql/trace.go, line 44 at r2](https://reviewable.io:443/reviews/cockroachdb/cockroach/19008#-KvZFVY_8ozvP4QZxsFF:-Kv_GWz_IZ7MoFj7YDKJ:b-896fix) ([raw file](https://github.com/cockroachdb/cockroach/blob/eca98ea2a80df677baa30f90cd5076d8bae76299/pkg/sql/trace.go#L44)):*
<details><summary><i>Previously, jordanlewis (Jordan Lewis) wrote…</i></summary><blockquote>

nit: i don't think this should be capitalized since it's not exported. Or is it?
</blockquote></details>

Done.

---

*[pkg/sql/trace.go, line 59 at r2](https://reviewable.io:443/reviews/cockroachdb/cockroach/19008#-KvZFZMyue5HLNvNiAkj:-Kv_H95i8H1aInlfeMyH:b-896fix) ([raw file](https://github.com/cockroachdb/cockroach/blob/eca98ea2a80df677baa30f90cd5076d8bae76299/pkg/sql/trace.go#L59)):*
<details><summary><i>Previously, jordanlewis (Jordan Lewis) wrote…</i></summary><blockquote>

nit: caps again.
</blockquote></details>

Done.

---

*[pkg/sql/sqlbase/kvfetcher.go, line 227 at r2](https://reviewable.io:443/reviews/cockroachdb/cockroach/19008#-KvZF0ROAeMSnNp1agq7:-Kv_Er6ueeVZHWqhpMEx:b79zxk1) ([raw file](https://github.com/cockroachdb/cockroach/blob/eca98ea2a80df677baa30f90cd5076d8bae76299/pkg/sql/sqlbase/kvfetcher.go#L227)):*
<details><summary><i>Previously, jordanlewis (Jordan Lewis) wrote…</i></summary><blockquote>

Hmmm.... not sure if this makes sense. This emits at least 1 Scan event per read operation. Isn't that pretty expensive?

In general it's not obvious to me that we should ever be doing this much unconditional tracing. But I guess that's another fish to fry - we already do a ton of unconditional tracing (including internals like time spent in the command queue). 
</blockquote></details>

The key here is "per read operation" - per batch sent for execution. These are big operations. Generally they're RPCs. Each one of them generates tons of messages already. One more here is not a big deal.

Unconditional logging is not particularly expensive when tracing and the respective logging is turned off - we've optimized this a bunch (when tracing/logging is enabled, that's another story). I believe we ran some tests about it. There must be some overhead for the boxing that the untyped `Eventf()` interface must do... We should test again at some point.

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/19008#-:-Kv_BcLL8N9Iqiy6PBuT:bmu3wiq)*
<!-- Sent from Reviewable.io -->