@andreimatei 2017-10-09
Added a benchmark and another commit changing a log message in traceNode.

---

Review status: 2 of 11 files reviewed at latest revision, 6 unresolved discussions, some commit checks pending.

---

*[pkg/sql/sqlbase/kvfetcher.go, line 227 at r8](https://reviewable.io:443/reviews/cockroachdb/cockroach/19008#-Kvn0VpjARpwX6gC0ImJ:-Kw1IPFRbVDYrCbxhNm4:b-v35ae4) ([raw file](https://github.com/cockroachdb/cockroach/blob/c2d7d3ee3281e2c0101ebd2aea9d1ea629a74ebe/pkg/sql/sqlbase/kvfetcher.go#L227)):*
<details><summary><i>Previously, andreimatei (Andrei Matei) wrote…</i></summary><blockquote>

[Pasting here from your email response. Please be aware that email responds are posted at top comments, so they break conversations.]

The fact we have too much overhead on the hot path is undeniable. This should be reduced, not increased.

> What else can I do for you?

A before/after benchmark comparison on an index join of 1000 rows, for example, counting both the time overhead and the allocations.
</blockquote></details>

I added a commit with the benchmark. No difference.
```
[andrei:~/work/src/github.com/cockroachdb/cockroach] sql-kv-tracing+ ± benchstat /tmp/{before,after}
name         old time/op  new time/op  delta
IndexJoin-8  8.54ms ± 4%  8.54ms ± 1%   ~     (p=0.605 n=9+9)


[andrei:~/work/src/github.com/cockroachdb/cockroach] sql-kv-tracing+ ± benchstat /tmp/{before-mem,after-mem}
name         old time/op    new time/op    delta
IndexJoin-8    8.64ms ± 2%    8.59ms ± 6%    ~     (p=0.243 n=9+10)

name         old alloc/op   new alloc/op   delta
IndexJoin-8    3.62MB ± 0%    3.62MB ± 0%  -0.03%  (p=0.000 n=9+9)

name         old allocs/op  new allocs/op  delta
IndexJoin-8     11.6k ± 0%     11.6k ± 0%    ~     (p=0.314 n=10+10)
```

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/19008#-:-Kw1IQgv0u7WrG0F4AVL:bdczo3b)*
<!-- Sent from Reviewable.io -->