@knz 2017-10-09
<img class="emoji" title=":lgtm:" alt=":lgtm:" align="absmiddle" src="https://reviewable.io/lgtm.png" height="20" width="61"/> modulo extending the commit message with the bench results

---

Reviewed 4 of 8 files at r10, 1 of 1 files at r11, 6 of 8 files at r12, 2 of 2 files at r13.
Review status: 11 of 12 files reviewed at latest revision, 6 unresolved discussions, all commit checks successful.

---

*[pkg/sql/sqlbase/kvfetcher.go, line 227 at r8](https://reviewable.io:443/reviews/cockroachdb/cockroach/19008#-Kvn0VpjARpwX6gC0ImJ:-Kw1moo7VgH3NdrkxMoQ:bnemfv1) ([raw file](https://github.com/cockroachdb/cockroach/blob/c2d7d3ee3281e2c0101ebd2aea9d1ea629a74ebe/pkg/sql/sqlbase/kvfetcher.go#L227)):*
<details><summary><i>Previously, andreimatei (Andrei Matei) wrote…</i></summary><blockquote>

I added a commit with the benchmark. No difference.
```
[andrei:~/work/src/github.com/cockroachdb/cockroach] sql-kv-tracing+ ± benchstat /tmp/{before,after}
name         old time/op  new time/op  delta
IndexJoin-8  8.54ms ± 4%  8.54ms ± 1%   ~     (p=0.605 n=9+9)


[andrei:~/work/src/github.com/cockroachdb/cockroach] sql-kv-tracing+ ± benchstat /tmp/{before-mem,after-mem}
name         old time/op    new time/op    delta
IndexJoin-8    8.64ms ± 2%    8.59ms ± 6%    ~     (p=0.243 n=9+10)

name         old alloc/op   new alloc/op   delta
IndexJoin-8    3.62MB ± 0%    3.62MB ± 0%  -0.03%  (p=0.000 n=9+9)

name         old allocs/op  new allocs/op  delta
IndexJoin-8     11.6k ± 0%     11.6k ± 0%    ~     (p=0.314 n=10+10)
```
</blockquote></details>

Brilliant, thanks. Please copy this into the commit message.

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/19008#-:-Kw1n3LUCBgyHgG8rn45:b-lnevib)*
<!-- Sent from Reviewable.io -->